Use requesting customer URI when interacting with DOIs (#2390)
We have observed a bug where DOIs are minted in the wrong prefix due to
the current implementation using the `publication::getPublisher` method
to extract the requesting customer. This used to work because of the
tight binding between the `publisher` and the `owner` of the
publication. Since we in practice stopped using `publisher` some time
ago, this logic no longer holds for cases where the requesting customer
for DOI-events is not the same as the content of
`publication::getPublisher`. Specifically, these cases are:

  - imported publications from Cristin (SIkt is the publisher)
- publications curated by > 1 customer, where the requesting customer is
not the owner

This PR:
  - Sources requesting customer URI from requestInfo
  - Adds requesting customer to interactions with DOIs

Not in this PR:
  - No validation of the requesting customer (i.e. the request may fail)