fix: Remove excessive logging (#2393)
Removes excessive logging from the handler to reduce noise and log size.

The previous implementation included the entire JSON string of both the
new and old image, plus the entire resource itself. This is replaced
with logging just the (minimal) input event, the type and topic of the
update event, and the resource identifier. Example log output for
successful invocations:

```
2025-11-04T13:34:17,924+01:00  INFO  n.u.n.p.e.h.e.ExpandDataEntriesHandler:186 - Reading update event from S3 for input: {"topic":"ignored","uri":"s3://ignoredForFakeS3Client/bb35d2cc-e60c-45c0-b0ae-8e5fff6d4176","timestamp":"2025-11-04T12:34:17.916081Z"}

2025-11-04T13:34:21,081+01:00  INFO  n.u.n.p.e.h.e.ExpandDataEntriesHandler:99 - Handling update event with action=MODIFY and topic=PublicationService.Resource.Update

2025-11-04T13:34:28,847+01:00  INFO  n.u.n.e.ResourceExpansionServiceImpl:136 - Expanding Resource: 019a4edc63f6-2fa16f77-1641-4f62-a0b1-99b61af73e8b
```

Instead of logging the complete documents on every invocation, we now
log the final JSON document at the point of failure and only on failure.