Add checking of dates in publication instance publisher uris (#2434)
This addresses the issues we have observed with publishing channel URIs,
but moves the focus from PublicationContext to PublicationInstance where
we find manifestations regretfully embedded due to a desire to reduce
modelling complexity….

The code probably has a few optimisations that are possible since code
is repeated in at least three places.

I have attempted to find and test every instance.