refactor: Use shared version catalog (#2438)
This is a first step towards using the shared version catalog (like we
do in `nva-nvi`) for dependencies. It touches many files and lines, but
the changes in `build.gradle` files are mostly "find and replace".

This doesn't add much value in itself, but if we upgrade a few
dependencies and do some minor refactoring, we can remove almost
everything from the local version catalog. This would make future
version upgrades much simpler.

Changes:

- Adds setup for using shared version catalog from `nva-commons`
- Replaces references to local catalog with identical items from the
shared catalog
- Removes unused items from the local version catalog